Proof of concept for not-masquerading namespaces.
This currently requires a special network namespace,
nonsense-temporary-upstream, to use as an upstream, and requires it to
be set up for NAT to have access to outside world, but hey, the subset
of things that are in right now actually _works_.

Also included a dummy `nonsensectl` version that is necessary for the
proof of concept.

Also fill the README in somewhat.