Make a more complicated example work.
* More namespaces with some sort of config in systemctl.
* Add a default nft configuration for router namespaces.
* Add slice configuration and drop-in files.
* Add units necessary to make the example work.

At this point in time, it seems that:
* networks work;
* routers work;
* clients work.

One thing that _may_ require further work related to restructoring the
service units is having a network with another network as an uplink,
which is a weird edge case that I have in my private example configuration
mockup on the side, that I definitely want to support, though I _think_
that this can be achieved by just starting the router service in the
uplink `net.*` namespace for as long as it has a downlink network
namespace.

The next brand new thing to achieve is taking control over interfaces
and the default namespace, but I think I need to start working on a
testing environment, probably with Vagrant as the sandbox for a kernel,
and then Docker as a way to be able to run the tests multiple times
without rebooting even if things explode while running a testcase.