Make network connections to external netnses work.
* Initial actual versions of nonsensectl and nonsensed. Further commits
should require smaller patches because of the groundwork provided here.
* Install all the things in correct places.
* Nuke the unnecessary network specific systemd units, fold them into
nonsense-connection@.service.
* Make nonsense-connection@.service use external scripts for the work it
needs to do (the last version that I had that did everything inside the
unit file was truly ridiculous).
* There is a partial documentation of the systemd interfaces in one of
the source files of the daemon; that needs to be filled in (with methods
of Config and MutableConfig) and moved to a proper, separate location in
the future.
* Both the controller and the daemon really need an IP address library.
I should fix something up sooner rather than later.
* Add initial network tests and make them work.