Massively refactor how nonsense deals with entities.
Before, nonsense knew of three kinds of entities: interfaces, networks,
and namespaces. I was growing increasingly certain that that was
misguided, especially so when I started considering how other (than net)
kinds of namespaces would fit into everything here, so I went back to
the drawing board.

As of this commit, nonsense knows of the concepts of "entities" and
"component". Think of an "entity" like you'd think of a systemd unit
(indeed, they spawn cgroup slices), and of a "component" like you'd
think of a namespace of a specific kind. An entity is composed of zero
or more components, and this defines what namespaces the processes
inside said entity will live in.

Entities have a hierarchy embedded in their naming scheme; they are
dot-delimited names, so foo.bar.baz and foo.buzz share one segment of
the naming. There will be features that use this directly; right now the
main thing it translates to is the structure of the cgroups generated.
foo.bar.baz creates a nonsense-foo-bar-baz.slice unit, while foo.buzz
creates a nonsense-foo-buzz.slice unit, using systemd's cgroups
interface.

There's now also an additional daemon binary, nonsense-entityd, which is
spawned as the main ("init") (sometimes literally pid 1, once I implement
PID namespace support in nonsense) process that _actually_ implements
all the necessary system mechanisms.

There's almost feature parity between this commit and the previous one;
the previous one had working router namespaces, this one doesn't. This
will be fixed soon. Otherwise, L2/L3 switches (new terminology for
"networks") work, and the tests are all passing.

There's a lot that will be enabled by this; to name one, delegation of
components to another entity to enable sharing certain kinds of
namespaces and the creation of mix-and-match entities. Currently it's a
bit of a mess, but this mess works and should be fairly easy to reduce
as nonsense moves forwards, so that's good.

One big thing that went away is the need for "property" getters for the
functioning of the namespaces, and of namespace locking. The properties
_may_ return, for the purpose of enabling users to write scripts, but
will probably require a good motivation before I delve into that messy
mechanism again. Similarly, locks are gone, because the entityd process
now always has an atomic snapshot of the world, and therefore doesn't
need to protect from changes to the configuration _while it's doing its
thing_. A return of properties would need to somehow fix this problem
again, which is one more reason I'm happy they are no more at this point
in time.

Finally, to ensure that operations on the same entity don't interleave
and for instance cause weird results due to settings changing mid-start
or some such, implement a simple per-entity-name queue mechanism that
all operations changing the configuration or state of one or more
entities are supposed to use, effectively creating additional ordering
on top of dbus message ordering.