Support foreign-architecture testing with Debian images.
Move the *-clang images to use lld as the linker. (This was necessary to
get clang builds working for Debian mips64el, but seems like a good idea
in general).

Make all gcc configurations, except for the latest Ubuntu LTS, install
`gcc` instead of `gcc-10` now that GCC 10 is the default version on all
of them.

Fix a stack overflow that would happen in the loop waiting for the
entityd dbus connection to finish initializing. I used to use
synchronous coroutine transfer there, which turns out to be a *very* bad
idea in a loop. This happened quite often on the Fedora images, and all
the time on the qemu-user-static-powered foreign images. Now it's using
the bool-returning form of `await_suspend` instead.